We shall run as root because reasons
We run as root so we can just stick it on port 22 without having to use
authbind (not availble on centos by default) or Linux capabilites not
availble in our environment because we are not actullay running on a
Linux Kernel. Were running cowrie inside a zone anyway so this
serves as our protection. Root in the zone is fine.